Move test case to it's own namespace
