+++ osu.Game.Rulesets.Karaoke.Tests/Objects/ToneCalculationTest.cs
@@ -4,7 +4,7 @@
 using NUnit.Framework;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Tests
+namespace osu.Game.Rulesets.Karaoke.Tests.Objects
 {
     [TestFixture]
     public class ToneCalculationTest
