+++ osu.Game.Rulesets.Karaoke.Tests/UI/TestCaseControlOverlay.cs
@@ -9,7 +9,7 @@ using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.UI;
 using osu.Game.Tests.Visual;
 
-namespace osu.Game.Rulesets.Karaoke.Tests
+namespace osu.Game.Rulesets.Karaoke.Tests.UI
 {
     [TestFixture]
     public class TestCaseControlOverlay : OsuTestScene
